Menu: tela inicial e criar conta, loop para criação de conta e vários IFs para evitar duplicatas

+++ classes/banco.py
@@ -1,10 +1,7 @@
 import sqlite3, os
 
 class Banco:
-    def __init__(self):
-        pass
-
-    def abrirConexao(self): ## INICIAR O BANCO
+    def __init__(self): ## INICIAR O BANCO
         self.banco = sqlite3.connect('usuarios.db')
         self.cursor = sqlite3.Cursor(self.banco)
     
@@ -20,25 +17,37 @@ class Banco:
         self.banco.commit()
 
     def criarUsuario(self, usuario): ## CRIAR USUARIO
-        self.cursor.execute("INSERT INTO User VALUES (NULL, '"+usuario.user+"', '"+usuario.password+"')")
-        self.banco.commit()
+        if self.verificarDados(usuario):
+            input('Usuario já cadastrado, tente novamente! Aperte qqr tecla para continuar.')
+        else:
+            self.cursor.execute("INSERT INTO User VALUES (NULL, '"+usuario.user+"', '"+usuario.password+"')")
+            self.banco.commit()
+            return True
 
     def lerUsuario(self, id): ## LER USUARIO
-        self.cursor.execute("SELECT * FROM User WHERE id = "+id+"")
-        r = self.cursor.fetchall()
+        if self.verificarId(id):
+            self.cursor.execute("SELECT * FROM User WHERE id = "+id+"")
+            r = self.cursor.fetchall()
 
-        return r[0]
+            return r[0]
+        else: input('ID não encontrado, tente novamente! Aperte qqr tecla para continuar.')
     
     def atualizarUsuario(self, id, b, valor): ## ATUALIZAR USUARIO
-        if b == 'U':
-            self.cursor.execute("UPDATE User SET nome = '"+valor+"' WHERE id = "+id+"")
-        elif b == 'S':
-            self.cursor.execute("UPDATE User SET senha = '"+valor+"' WHERE id = "+id+"")
-        self.banco.commit()
+        if self.verificarId(id):
+            if b == 'U':
+                self.cursor.execute("UPDATE User SET nome = '"+valor+"' WHERE id = "+id+"")
+            elif b == 'S':
+                self.cursor.execute("UPDATE User SET senha = '"+valor+"' WHERE id = "+id+"")
+            self.banco.commit()
+            return True
+        else: input('ID não encontrado, tente novamente! Aperte qqr tecla para continuar.')
     
     def deletarUsuario(self, id): ## DELETAR USUARIO
-        self.cursor.execute("DELETE FROM User WHERE id = "+id+"")
-        self.banco.commit()
+        if self.verificarId(id):
+            self.cursor.execute("DELETE FROM User WHERE id = "+id+"")
+            self.banco.commit()
+            return True
+        else: input('ID não encontrado, tente novamente! Aperte qqr tecla para continuar.')
     
     ## VERIFICAR SE HÁ DADOS COMPATIVEIS NO BANCO
 
@@ -47,5 +56,13 @@ class Banco:
         r = self.cursor.fetchall()
 
         for v in r:
-            if v[1] == usuario.user and v[2] == usuario.password:
+            if (v[1] == usuario.user and v[2] == usuario.password) or v[1] == usuario.user:
+                return True
+    
+    def verificarId(self, id):
+        self.cursor.execute("SELECT * FROM User")
+        r = self.cursor.fetchall()
+
+        for v in r:
+            if v[0] == id:
                 return True
